куліковський-георгій/rest_file/requirements.txt
Django == 4.2.4
djangorestframework
django-cors-headers
djangorestframework-simplejwt